refactor: 일부 코드 수정 및 불필요한 코드 삭제
