+++ src/main/java/com/sprint/mission/discodeit/service/AuthService.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.service;
 
 import com.sprint.mission.discodeit.entity.User;
-import com.sprint.mission.discodeit.service.dto.AuthServiceLoginDto;
+import com.sprint.mission.discodeit.service.dto.authdto.AuthServiceLoginDto;
 
 public interface AuthService {
