+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicAuthService.java
@@ -3,7 +3,7 @@ package com.sprint.mission.discodeit.service.basic;
 import com.sprint.mission.discodeit.entity.User;
 import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.AuthService;
-import com.sprint.mission.discodeit.service.dto.AuthServiceLoginDto;
+import com.sprint.mission.discodeit.service.dto.authdto.AuthServiceLoginDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
